prediction/requirements-ml.txt
# HYDRA-LB Prediction Module - ML Dependencies
# 
# Install with: pip install -r requirements-ml.txt
# Or with uv: uv pip install -r requirements-ml.txt
#
# This file is kept separate from the main requirements.txt
# for portability to systems without the full ML environment.

# Core ML
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Data processing
scikit-learn>=1.3.0

# Visualization & Logging
tensorboard>=2.14.0
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0
